# generated/types/pillarDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Mishkat Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PillarDetailLinksItem } from "./pillarDetailLinksItem";
import type { VideoItem } from "./videoItem";

export interface PillarDetail {
  id: string;
  name: string;
  arabicName: string;
  order: number;
  description: string;
  arabicText: string;
  videos: VideoItem[];
  links: PillarDetailLinksItem[];
}