# generated/types/quizResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Mishkat Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizResultFeedbackItem } from "./quizResultFeedbackItem";

export interface QuizResult {
  quizId: string;
  score: number;
  totalQuestions: number;
  percentage: number;
  passed: boolean;
  feedback: QuizResultFeedbackItem[];
}