# generated/types/videoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Mishkat Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoItemType } from "./videoItemType";

export interface VideoItem {
  id: string;
  title: string;
  description: string;
  type: VideoItemType;
  duration: string;
  thumbnailUrl: string;
  videoUrl: string;
  subjectId: string;
  uploadedAt: string;
}