# platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Mishkat Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalVideos: number;
  totalPdfs: number;
  totalImages: number;
  totalQuizzes: number;
  totalSubjects: number;
  totalPillars: number;
}